Fix #build_checkout_request_body
In commit https://github.com/nebulab/solidus_viabill/pull/15/commits/7e1440811db87ad527315633791f8901f19b4a33 we changed request_body[:success_url] for gateway.success_url
which generates an error because the former returns a url with params, while the latter without.
With this commit we are reverting that change and aligning the rest of the parameters to match
in format with this one.